Improve Robot's move exit condition
It is not the job of the move method to check for those.
isTooClose() is especially bad, as it does not allow the robot to turn
away from a wall. Checking if the target cell is available, should be
done by the path finder, or other logic. If collision detetion is desired,
then the move method can check for a COLLISION boolean variable, and if true,
pass over movement control to a collision avoidance logic.